requirements.txt
numpy~=2.2.1
pillow~=11.1.0
scipy~=1.15.1